MAINT: Move model calibration purely to the main model calibration file.
